requirements.txt
streamlit
faiss-cpu
langchain
pandas
numpy
sentence-transformers
